Items/pyproject.toml
[project]
name = "items"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents[litellm]>=0.1.0",
]
